Delete old duplicate ChainedResourceTransformer. Break link between ChainedResourceTransformationContext and ResourceTransformationContext to avoid the 'illegal' methods
